requirements.txt
langchain
openai
pinecone-client
langchain-pinecone
langchain-openai
python-dotenv
pypdf
sentence-transformers
langchain-huggingface 
text-generation
transformers 
google-search-results 
numexpr 
langchainhub
sentencepiece
jinja2
bitsandbytes
accelerate
langchain-together